remove unneccessary files and add code for webhook functionality in single file
